v3.2.2 - start to refactor
