+++ README.md
@@ -1,0 +1,20 @@
+# Enchant Outline ( for 1.21.1 NeoForge )
+
+> Unofficial NeoForge port for 1.21.1
+> Because I wanted it
+
+## Special Thanks
+
+The original idea from the original repo, I just forked it and used another method.
+
+## About
+
+I'm not writing one, go read it yourself
+
+## Progress Map
+
+- [x] 1.21.1 NeoForge
+- [x] Item rendering ( Shader + Mixin )
+- [ ] Armor rendering
+- [ ] Trident rendering
+- [ ] Proper tests ( sodium compat unknown / performance? / all items )
