+++ gradle.properties
@@ -1,15 +1,11 @@
-# Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
 minecraft_version=1.21.1
 neoforge_version=21.1.230
 
-yacl_version=3.8.1+1.21.1-neoforge
-sodium_version=tVLDCyrm
-
 # Mod Properties
-mod_version=3.2
+mod_version=3.2.1
 maven_group=net.enchantoutline
 archives_base_name=enchant_outline
