+++ src/main/java/net/enchantoutline/EnchantmentGlintOutline.java
@@ -1,7 +1,7 @@
 package net.enchantoutline;
 
-import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -9,8 +9,4 @@ import org.slf4j.LoggerFactory;
 public class EnchantmentGlintOutline {
     public static final String MOD_ID = "enchant_outline";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
-
-
-    public EnchantmentGlintOutline(IEventBus modEventBus) {
-    }
 }
