+++ src/main/java/net/enchantoutline/ShaderRegistry.java
@@ -1,17 +1,16 @@
-package net.enchantoutline.events;
+package net.enchantoutline;
 
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
-import net.enchantoutline.EnchantmentGlintOutline;
-import net.enchantoutline.util.Shaders;
-import net.minecraft.client.renderer.ShaderInstance;
-import net.minecraft.resources.ResourceLocation;
-import net.neoforged.api.distmarker.Dist;
+
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.RegisterShadersEvent;
 
-@EventBusSubscriber(modid = EnchantmentGlintOutline.MOD_ID, value = Dist.CLIENT)
-public class ClientModEvent {
+import net.minecraft.client.renderer.ShaderInstance;
+import net.minecraft.resources.ResourceLocation;
+
+@EventBusSubscriber(modid = EnchantmentGlintOutline.MOD_ID)
+public class ShaderRegistry {
 
     @SubscribeEvent
     public static void onRegisterShaders(RegisterShadersEvent event) {
