+++ src/main/java/net/enchantoutline/Shaders.java
@@ -1,7 +1,8 @@
-package net.enchantoutline.util;
+package net.enchantoutline;
 
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
 import com.mojang.blaze3d.vertex.VertexFormat;
+
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.RenderStateShard;
 import net.minecraft.client.renderer.RenderType;
