+++ src/main/java/net/enchantoutline/mixin/HumanoidArmorLayerMixin.java
@@ -1,0 +1,36 @@
+package net.enchantoutline.mixin;
+
+import net.enchantoutline.Shaders;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+
+import net.minecraft.client.model.HumanoidModel;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.entity.layers.HumanoidArmorLayer;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+
+@Mixin(HumanoidArmorLayer.class)
+public class HumanoidArmorLayerMixin {
+
+    @Inject(method = "renderArmorPiece(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/world/entity/EquipmentSlot;ILnet/minecraft/client/model/HumanoidModel;)V", at = @At("HEAD"))
+    private void addArmorOutlinePass(PoseStack poseStack, MultiBufferSource bufferSource, LivingEntity entity, EquipmentSlot slot, int light, HumanoidModel<?> model, CallbackInfo ci) {
+        net.minecraft.world.item.ItemStack armorStack = entity.getItemBySlot(slot);
+        if (armorStack.isEmpty() || !armorStack.isEnchanted()) return;
+
+        VertexConsumer outlineBuffer = bufferSource.getBuffer(Shaders.getOutlineLayer());
+        poseStack.pushPose();
+
+        float scale = 1.05f;
+        poseStack.scale(scale, scale, scale);
+        model.renderToBuffer(poseStack, outlineBuffer, light, net.minecraft.client.renderer.texture.OverlayTexture.NO_OVERLAY);
+        poseStack.popPose();
+    }
+
+}
