--- src/main/resources/META-INF/accesstransformer.cfg
@@ -1,13 +1,0 @@
-public net.minecraft.client.model.geom.ModelPart f_104274_
-public net.minecraft.client.model.geom.ModelPart f_104275_
-public-f net.minecraft.client.model.geom.ModelPart$Cube f_104297_
-
-public net.minecraft.client.renderer.RenderType$CompositeRenderType
-public net.minecraft.client.renderer.RenderType$CompositeState
-public net.minecraft.client.renderer.RenderType$CompositeRenderType state
-public net.minecraft.client.renderer.RenderType$CompositeState cullState
-public net.minecraft.client.renderer.RenderType$CompositeState textureState
-public net.minecraft.client.renderer.RenderStateShard$TextureStateShard texture
-
-public net.minecraft.client.renderer.rendertype.RenderSetup$TextureBinding
-public net.minecraft.client.renderer.rendertype.RenderType <init>(Ljava/lang/String;Lnet/minecraft/client/renderer/rendertype/RenderSetup;)V
