+++ src/main/resources/assets/enchant_outline/shaders/core/outline.json
@@ -21,8 +21,7 @@
         { "name": "ModelViewMat",  "type": "matrix4x4", "count": 16, "values": [ 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 ] },
         { "name": "ProjMat",       "type": "matrix4x4", "count": 16, "values": [ 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 ] },
         { "name": "TextureMat",    "type": "matrix4x4", "count": 16, "values": [ 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 ] },
-        { "name": "ColorModulator","type": "float",     "count": 4,  "values": [ 1.0, 1.0, 1.0, 1.0 ] },
-        { "name": "GlintAlpha",    "type": "float",     "count": 1,  "values": [ 1.0 ] }
+        { "name": "ColorModulator","type": "float",     "count": 4,  "values": [ 1.0, 1.0, 1.0, 1.0 ] }
     ],
     "samplers": [
         { "name": "Sampler0" }
