+++ src/main/resources/enchant_outline.mixins.json
@@ -9,6 +9,7 @@
 		"requireAnnotations": true
 	},
 	"client": [
+		"HumanoidArmorLayerMixin",
 		"ItemRendererMixin"
 	]
 }
